[INDY-1992] move tests to indy_node/tests/package
