+++ indy_node/test/package/test_metadata.py
@@ -1,7 +1,7 @@
 import pytest
 
 from indy_node.__metadata__ import (
-    check_version, set_version, load_version, pep440_version
+    check_version, set_version, load_version, pep440_version, split_version_from_str
 )
 
 
@@ -55,3 +55,20 @@ def test_pep440_version_dev():
 
 def test_pep440_version_rc():
     pep440_version((1, 2, 3, 'rc', 2)) == "1.2.3.rc2"
+
+
+def test_split_version_from_str_only_int():
+    assert split_version_from_str("1.2.3.4.5") == [1, 2, 3, 4, 5]
+
+
+def test_split_version_from_str_with_str():
+    assert split_version_from_str("1.2.3.rc.5") == [1, 2, 3, 'rc', 5]
+
+
+def test_check_version_right_case():
+    check_version(split_version_from_str("1.2.3.rc.5"))
+
+
+def test_check_version_wrong_case():
+    with pytest.raises(ValueError):
+        check_version(split_version_from_str("1.2.3.4.5"))
